requirements.txt
onnxruntime
urllib3
tf2onnx
keras
mxnet
numpy
coremltools
onnxmltools
